[UMA] moved UMA _template file to apps
