+++ apps/uma/_template/conv2dnchw.cc
